requirements.txt
# OpenAI API
openai>=1.12.0

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Text matching and search
rapidfuzz>=3.6.0
rank-bm25>=0.2.2

# HTTP requests (if needed)
requests>=2.31.0

